Backend/requirements.txt
Django==1.7
Pillow==2.6.1
argparse==1.2.1
boto==2.34.0
dj-database-url==0.3.0
dj-static==0.0.6
django-extensions==1.4.6
django-registration-redux==1.1
django-storages==1.1.8
django-toolbelt==0.0.1
djangorestframework==2.4.3
gunicorn==19.1.1
ipython==2.3.0
psycopg2==2.5.4
qrcode==5.1
selenium==2.44.0
six==1.8.0
static3==0.5.1
whitenoise==1.0.3
wsgiref==0.1.2
